update du main, add_coord, libit.h

+++ includes/libit.h
@@ -6,7 +6,7 @@
 /*   By: stmartin <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/12/22 17:45:57 by stmartin          #+#    #+#             */
-/*   Updated: 2015/12/23 19:06:56 by stmartin         ###   ########.fr       */
+/*   Updated: 2015/12/26 16:23:18 by stmartin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,7 @@
 
 # include "libft.h"
 # include <fcntl.h>
+#include <stdlib.h>
 
 typedef struct		s_coord
 {
@@ -24,6 +25,6 @@ typedef struct		s_coord
 
 int					checkbuff(char *buff);
 int					check_valid(char *buff);
-t_coord				**add_coord(char *buff, int i, int x, int y);
+t_coord				**add_coord(char *buff, size_t i, int x, int y);
 char				*fill_buff(int fd);
 #endif
